requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2
tqdm==4.66.4
selenium==4.24.0
webdriver-manager==4.0.1
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
requests==2.32.3
